вынес из MainActivity.kt классы и интерфейсы.
