+++ app/src/main/java/com/bignerdranch/android/stopwatch/MainActivity.kt
@@ -1,13 +1,14 @@
 package com.bignerdranch.android.stopwatch
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.TextView
-import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
+import androidx.appcompat.app.AppCompatActivity
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
 
@@ -58,157 +59,15 @@ class MainActivity : AppCompatActivity() {
     }
 }
 
-sealed class StopwatchState {
 
-    data class Paused(
-        val elapsedTime: Long
-    ) : StopwatchState()
 
-    data class Running(
-        val startTime: Long,
-        val elapsedTime: Long
-    ) : StopwatchState()
-}
 
-interface TimestampProvider {
-    fun getMilliseconds(): Long
-}
 
-class StopwatchStateCalculator(
-    private val timestampProvider: TimestampProvider,
-    private val elapsedTimeCalculator: ElapsedTimeCalculator,
-) {
 
-    fun calculateRunningState(oldState: StopwatchState): StopwatchState.Running =
-        when (oldState) {
-            is StopwatchState.Running -> oldState
-            is StopwatchState.Paused -> {
-                StopwatchState.Running(
-                    startTime = timestampProvider.getMilliseconds(),
-                    elapsedTime = oldState.elapsedTime
-                )
-            }
-        }
 
-    fun calculatePausedState(oldState: StopwatchState): StopwatchState.Paused =
-        when (oldState) {
-            is StopwatchState.Running -> {
-                val elapsedTime = elapsedTimeCalculator.calculate(oldState)
-                StopwatchState.Paused(elapsedTime = elapsedTime)
-            }
-            is StopwatchState.Paused -> oldState
-        }
-}
 
-class ElapsedTimeCalculator(
-    private val timestampProvider: TimestampProvider,
-) {
 
-    fun calculate(state: StopwatchState.Running): Long {
-        val currentTimestamp = timestampProvider.getMilliseconds()
-        val timePassedSinceStart = if (currentTimestamp > state.startTime) {
-            currentTimestamp - state.startTime
-        } else {
-            0
-        }
-        return timePassedSinceStart + state.elapsedTime
-    }
-}
 
-class TimestampMillisecondsFormatter {
 
-    fun format(timestamp: Long): String {
-        val millisecondsFormatted = (timestamp % 1000).pad(3)
-        val seconds = timestamp / 1000
-        val secondsFormatted = (seconds % 60).pad(2)
-        val minutes = seconds / 60
-        val minutesFormatted = (minutes % 60).pad(2)
-        val hours = minutes / 60
-        return if (hours > 0) {
-            val hoursFormatted = (minutes / 60).pad(2)
-            "$hoursFormatted:$minutesFormatted:$secondsFormatted"
-        } else {
-            "$minutesFormatted:$secondsFormatted:$millisecondsFormatted"
-        }
-    }
 
-    private fun Long.pad(desiredLength: Int) = this.toString().padStart(desiredLength, '0')
 
-    companion object {
-        const val DEFAULT_TIME = "00:00:000"
-    }
-}
-
-class StopwatchStateHolder(
-    private val stopwatchStateCalculator: StopwatchStateCalculator,
-    private val elapsedTimeCalculator: ElapsedTimeCalculator,
-    private val timestampMillisecondsFormatter: TimestampMillisecondsFormatter
-) {
-
-    var currentState: StopwatchState = StopwatchState.Paused(0)
-        private set
-
-    fun start() {
-        currentState = stopwatchStateCalculator.calculateRunningState(currentState)
-    }
-
-    fun pause() {
-        currentState = stopwatchStateCalculator.calculatePausedState(currentState)
-    }
-
-    fun stop() {
-        currentState = StopwatchState.Paused(0)
-    }
-
-    fun getStringTimeRepresentation(): String {
-        val elapsedTime = when (val currentState = currentState) {
-            is StopwatchState.Paused -> currentState.elapsedTime
-            is StopwatchState.Running -> elapsedTimeCalculator.calculate(currentState)
-        }
-        return timestampMillisecondsFormatter.format(elapsedTime)
-    }
-}
-
-class StopwatchListOrchestrator(
-    private val stopwatchStateHolder: StopwatchStateHolder,
-    private val scope: CoroutineScope,
-) {
-
-    private var job: Job? = null
-    private val mutableTicker = MutableStateFlow("")
-    val ticker: StateFlow<String> = mutableTicker
-
-    fun start() {
-        if (job == null) startJob()
-        stopwatchStateHolder.start()
-    }
-
-    private fun startJob() {
-        scope.launch {
-            while (isActive) {
-                mutableTicker.value = stopwatchStateHolder.getStringTimeRepresentation()
-                delay(20)
-            }
-        }
-    }
-
-    fun pause() {
-        stopwatchStateHolder.pause()
-        stopJob()
-    }
-
-    fun stop() {
-        stopwatchStateHolder.stop()
-        stopJob()
-        clearValue()
-    }
-
-    private fun stopJob() {
-        scope.coroutineContext.cancelChildren()
-        job = null
-    }
-
-    private fun clearValue() {
-        mutableTicker.value = "00:00:000"
-    }
-}
