+++ app/src/main/java/com/bignerdranch/android/stopwatch/StopwatchState.kt
@@ -1,0 +1,13 @@
+package com.bignerdranch.android.stopwatch
+
+sealed class StopwatchState {
+
+    data class Paused(
+        val elapsedTime: Long
+    ) : StopwatchState()
+
+    data class Running(
+        val startTime: Long,
+        val elapsedTime: Long
+    ) : StopwatchState()
+}
