+++ app/src/main/java/com/bignerdranch/android/stopwatch/TimestampProvider.kt
@@ -1,0 +1,5 @@
+package com.bignerdranch.android.stopwatch
+
+interface TimestampProvider {
+    fun getMilliseconds(): Long
+}
